Simplified error handling
