+++ include/readout/FrameErrorRegistry.hpp
@@ -8,15 +8,32 @@
 #ifndef READOUT_INCLUDE_READOUT_FRAMEERRORREGISTRY_HPP_
 #define READOUT_INCLUDE_READOUT_FRAMEERRORREGISTRY_HPP_
 
-#include <vector>
-#include <queue>
-#include <readout/FrameError.hpp>
+#include <deque>
 
 namespace dunedaq {
   namespace readout {
 
     class FrameErrorRegistry {
     public:
+      struct FrameError {
+      public:
+        FrameError(uint64_t start_ts, uint64_t end_ts) : start_ts(start_ts), end_ts(end_ts)
+        {
+
+        }
+
+        uint64_t start_ts;
+        uint64_t end_ts;
+
+        bool operator<(const FrameError& other) const {
+          return end_ts < other.end_ts;
+        }
+
+        bool operator>(const FrameError& other) const {
+          return end_ts > other.end_ts;
+        }
+      };
+
       FrameErrorRegistry()
       : m_errors()
       {
@@ -24,12 +41,12 @@ namespace dunedaq {
       }
 
       void add_error(FrameError error) {
-        m_errors.push(error);
+        m_errors.push_back(error);
       }
 
       void update_latest_frame_in_buffer(uint64_t ts) {
-        while (has_error() && m_errors.top().end_ts < ts) {
-          m_errors.pop();
+        while (has_error() && m_errors.front().end_ts < ts) {
+          m_errors.pop_front();
         }
       }
 
@@ -39,7 +56,7 @@ namespace dunedaq {
 
 
     private:
-      std::priority_queue<FrameError, std::vector<FrameError>, std::greater<FrameError>> m_errors;
+      std::deque<FrameError> m_errors;
     };
 
   } // namespace readout
